pyproject.toml
[tool.black]
line-length = 120
target_version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | \.vscode
  | postgres_data
)
'''

[tool.isort]
profile = "black"

[tool.poetry]
name = "libupd8"
version = "0.1.0"
description = ""
authors = ["Alexander Shishenko <alex@shishenko.com>"]

[tool.poetry.dependencies]
python = "^3.9"
# Core components
Django = "~3.2.5"
celery = {extras = ["redis"], version = "~5.1"}
# Other components
httpx = "^0.19.0"
lxml = "^4.6.3"
packaging = "^21.0"
psycopg2 = "^2.9.1"
python-telegram-bot = "^13.7"
emoji = "^1.2.0"
django-celery-beat = "^2.2.1"
django-model-utils = "^4.1.1"
sentry-sdk = "^1.3.0"
whitenoise = "^5.3.0"
Hypercorn = "^0.11.2"
python-dotenv = "^0.19.0"
django-environ = "^0.4.5"
django-redis = "^5.0.0"
tqdm = "^4.62.2"
semver = "^2.13.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
ipython = "^7.27.0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
